% Generated by roxygen2 (4.0.2): do not edit by hand
\name{slick.sql}
\alias{slick.sql}
\title{Evaluates given sql template against created session}
\usage{
slick.sql(session, sql, ...)
}
\arguments{
\item{sql}{is a template. \{\{id\}\} would be replaced by provided parameter with name `id`}

\item{...}{list of named parameters}
}
\description{
Evaluates given sql template against created session
}

