% Generated by roxygen2 (4.0.2): do not edit by hand
\name{slick.session}
\alias{slick.session}
\title{Creates RSlick session that should be provided to slick.sql}
\usage{
slick.session(url, driver, user, password)
}
\arguments{
\item{url}{URL to database}

\item{driver}{JDBC driver to be used for connection}

\item{user}{User name that has access to database}

\item{password}{User password that has access to database}
}
\description{
Creates RSlick session that should be provided to slick.sql
}

